requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.3
djangorestframework==3.6.3
flake8==3.3.0
greenlet==0.4.12
gunicorn==19.7.1
httpie==0.9.9
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
mccabe==0.6.1
msgpack-python==0.4.8
neovim==0.1.13
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.5.0
Pygments==2.2.0
pytz==2017.2
requests==2.18.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
virtualfish==1.0.5
wcwidth==0.1.7
yapf==0.16.3
